# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Debt Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /usaha:
    get:
      operationId: getUsahaList
      tags: [usaha]
      summary: Get all businesses (super admin)
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usaha"
    post:
      operationId: createUsaha
      tags: [usaha]
      summary: Create a business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUsahaBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usaha"
  /usaha/{id}:
    get:
      operationId: getUsaha
      tags: [usaha]
      summary: Get a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usaha"
    put:
      operationId: updateUsaha
      tags: [usaha]
      summary: Update a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUsahaBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usaha"
...
  /pelanggan:
    get:
      operationId: getPelangganList
      tags: [pelanggan]
      summary: Get all customers for owner's business
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pelanggan"
    post:
      operationId: createPelanggan
      tags: [pelanggan]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePelangganBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pelanggan"
  /pelanggan/{id}:
    get:
      operationId: getPelanggan
      tags: [pelanggan]
      summary: Get a customer with debt summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PelangganDetail"
    put:
      operationId: updatePelanggan
      tags: [pelanggan]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePelangganBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pelanggan"
    delete:
      operationId: deletePelanggan
      tags: [pelanggan]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /pembayaran:
    get:
      operationId: getPembayaranList
      tags: [pembayaran]
      summary: Get all payments for owner's business
      parameters:
        - name: hutang_id
          in: query
          schema:
            type: integer
        - name: pelanggan_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pembayaran"
    post:
      operationId: createPembayaran
      tags: [pembayaran]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePembayaranBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pembayaran"
  /pembayaran/{id}:
    delete:
      operationId: deletePembayaran
      tags: [pembayaran]
      summary: Delete a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Usaha:
      type: object
      properties:
        id:
          type: integer
        nama_usaha:
          type: string
        alamat:
          type: string
          nullable: true
        telepon:
          type: string
          nullable: true
        catatan:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, nama_usaha, created_at]
    CreateUsahaBody:
      type: object
      properties:
        nama_usaha:
          type: string
        alamat:
          type: string
          nullable: true
        telepon:
          type: string
          nullable: true
        catatan:
          type: string
          nullable: true
      required: [nama_usaha]
    Pelanggan:
      type: object
      properties:
        id:
          type: integer
        usaha_id:
          type: integer
        nama:
          type: string
        telepon:
          type: string
          nullable: true
        alamat:
          type: string
          nullable: true
        catatan:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, usaha_id, nama, created_at]
    PelangganDetail:
      type: object
      properties:
        id:
          type: integer
        usaha_id:
          type: integer
        nama:
          type: string
        telepon:
          type: string
          nullable: true
        alamat:
          type: string
          nullable: true
        catatan:
          type: string
          nullable: true
        created_at:
          type: string
        hutang_list:
          type: array
          items:
            $ref: "#/components/schemas/Hutang"
        pembayaran_list:
          type: array
          items:
            $ref: "#/components/schemas/Pembayaran"
        total_hutang:
          type: number
        total_dibayar:
          type: number
        sisa_hutang:
          type: number
      required: [id, usaha_id, nama, created_at, hutang_list, pembayaran_list, total_hutang, total_dibayar, sisa_hutang]
    CreatePelangganBody:
      type: object
      properties:
        nama:
          type: string
        telepon:
          type: string
          nullable: true
        alamat:
          type: string
          nullable: true
        catatan:
          type: string
          nullable: true
      required: [nama]
    Hutang:
      type: object
      properties:
        id:
          type: integer
        usaha_id:
          type: integer
        pelanggan_id:
          type: integer
        pelanggan_nama:
          type: string
        tanggal_hutang:
          type: string
        keterangan:
          type: string
          nullable: true
        nominal_hutang:
          type: number
        total_dibayar:
          type: number
        sisa_hutang:
          type: number
        status:
          type: string
          enum: [aktif, lunas]
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, usaha_id, pelanggan_id, pelanggan_nama, tanggal_hutang, nominal_hutang, total_dibayar, sisa_hutang, status, created_at, updated_at]
...
    Pembayaran:
      type: object
      properties:
        id:
          type: integer
        usaha_id:
          type: integer
        hutang_id:
          type: integer
        pelanggan_id:
          type: integer
        pelanggan_nama:
          type: string
        tanggal_bayar:
          type: string
        nominal_bayar:
          type: number
        catatan:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, usaha_id, hutang_id, pelanggan_id, pelanggan_nama, tanggal_bayar, nominal_bayar, created_at]
    CreatePembayaranBody:
      type: object
      properties:
        hutang_id:
          type: integer
        tanggal_bayar:
          type: string
        nominal_bayar:
          type: number
        catatan:
          type: string
          nullable: true
      required: [hutang_id, tanggal_bayar, nominal_bayar]
...