# lib/api-zod/src/generated/types/createPelangganBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Debt Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePelangganBody {
  nama: string;
  telepon?: string | null;
  alamat?: string | null;
  catatan?: string | null;
}