# lib/api-zod/src/generated/types/pelangganDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Debt Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Hutang } from "./hutang";
import type { Pembayaran } from "./pembayaran";

export interface PelangganDetail {
  id: number;
  usaha_id: number;
  nama: string;
  telepon?: string | null;
  alamat?: string | null;
  catatan?: string | null;
  created_at: string;
  hutang_list: Hutang[];
  pembayaran_list: Pembayaran[];
  total_hutang: number;
  total_dibayar: number;
  sisa_hutang: number;
}